feature/easy-api-platform-error-builder-7.0-cleanup
