+++ packages/EasyApiPlatform/src/EasyErrorHandler/Builder/CustomSerializerExceptionErrorResponseBuilder.php
@@ -3,31 +3,24 @@ declare(strict_types=1);
 
 namespace EonX\EasyApiPlatform\EasyErrorHandler\Builder;
 
-use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\Serializer\Exception\NotNormalizableValueException;
 use Symfony\Contracts\Service\Attribute\Required;
 use Throwable;
 
-final class ApiPlatformCustomSerializerExceptionErrorResponseBuilder extends
-    AbstractApiPlatformSerializerExceptionErrorResponseBuilder
+final class CustomSerializerExceptionErrorResponseBuilder extends AbstractDeserializationErrorResponseBuilder
 {
+    /**
+     * @var array<array{class: string, message_pattern: string, violation_message: string}>
+     */
     private array $customSerializerExceptions = [];
 
     /**
-     * @deprecated Deprecated since 6.4.0, will be moved to the parent class in 7.0
+     * @param array<array{class: string, message_pattern: string, violation_message: string}> $customSerializerExceptions
      */
-    private RequestStack $requestStack;
-
     #[Required]
     public function setCustomSerializerExceptions(array $customSerializerExceptions): void
     {
         $this->customSerializerExceptions = $customSerializerExceptions;
-    }
-
-    #[Required]
-    public function setRequestStack(RequestStack $requestStack): void
-    {
-        $this->requestStack = $requestStack;
     }
 
     protected function doBuildViolations(Throwable $throwable): array
@@ -55,29 +48,5 @@ final class ApiPlatformCustomSerializerExceptionErrorResponseBuilder extends
         }
 
         return [];
-    }
-
-    /**
-     * @param class-string|null $class
-     *
-     * @deprecated Deprecated since 6.4.0, will be moved to the parent class in 7.0
-     */
-    protected function normalizePropertyName(string $name, ?string $class = null): string
-    {
-        if ($class === null) {
-            $mainRequest = $this->requestStack->getMainRequest();
-
-            if ($mainRequest !== null) {
-                /** @var class-string|null $apiResourceClass */
-                $apiResourceClass = $mainRequest->attributes->get('_api_resource_class');
-                $class = $apiResourceClass;
-            }
-        }
-
-        if ($class !== null) {
-            return $this->nameConverter->normalize($name, $class);
-        }
-
-        return $name;
     }
 }
